backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-groq
tavily-python
datasets
pandas
python-dotenv
